Pointers in cpp
## Pointers

A pointer is used in C (and C++) to point to a memory location. If you'd like to know more about Memory Locations, read about the C Memory Model [here](https://www.notion.so/C-Concepts-eaf8eb8889714f4191abbeacfc765546).

Let's make a Pointer point to a variable `i` with a value of 4.

```cpp
int i = 4;
int *ptr = &i; 
```

The `&` is the **referencing** or **address** operator indicates that you are referring to the address of that location. The `*` is called the **dereferencing** operator and indicates that you are referring explicitly to the value and not the address. 

So, the above lines of code made a pointer, which points to the a variable `i` in the memory location (eg. 0x04) So, the pointer memory location contains the address of the variable `i` (0x04).

## Difference between * and &

Let's see the output of these various valid combinations of * and & on the pointer and the variable i. You can print a pointer using the `%p` format specifier.

```cpp
int i = 4;
int *ptr = &i;

    printf("Value at ptr = %p \n",ptr); 
    printf("Value at *ptr = %d \n",*ptr); 
    printf("Value at &ptr = %p \n", &ptr);
    printf("Value at i = %d \n",i); 
    printf("Value at &i = %p \n",&i);
```

The output is:

```
Value at ptr = 0x7ffdec27308c 
Value at *ptr = 4 
Value at &ptr = 0x7ffdec273090 
Value at i = 4 
Value at &i = 0x7ffdec27308c 
[Finished in 0.43s]
```