Classes Example using Cars
# Classes and Objects

Class is like a skeleton or blueprint for an object. It is a user defined data type, and many instances of the same class can be created containing copies of its members, called an object. 

A class contains data members and member functions. For example, for Student: 

- Data Members, eg. Student ID, Name of the Student, Age, etc.
- Member Functions (Method), eg. Apply for Sick Leave, Enrol for a class.

No storage is assigned for a class until an object is created.

```cpp
class ClassName
{
accessSpecifier: //either public, private or protected
ClassName() //Constructor
{
//initializations
}
memberFunction1(){}
};
```

Remember that Class definitions go on top of your main function! :D (if not in their own header file) 

Objects are instances of a class, and a special class function called **Constructor** performs the initialization of an object of a class. Similarly, a **Destructor** is used to destroy an Object. Constructors will have the same name as the Class Name, and no return value.

When an Object is created, it is initialized with the values that are declared from the Constructor function inside the class. If no Constructor is defined, the compiler calls the **Default Constructor** which initializes the object randomly. 

Here, I use a Car example to demonstrate how classes work in cpp.